:recycle: Split into libs and chat
